requirements.txt
aioredis==1.3.1
amqp==2.3.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.11.29
cffi==1.12.2
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.2
djangorestframework-simplejwt==3.3
dnspython==1.16.0
eventlet==0.24.1
gevent==1.4.0
greenlet==0.4.15
hiredis==1.1.0
hyperlink==20.0.1
idna==2.7
incremental==17.5.0
kombu==4.2.1
monotonic==1.5
msgpack==1.0.0
numpy==1.15.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
pytz==2018.7
redis==2.10.6
requests==2.20.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.24.1
vine==1.1.4
zope.interface==5.2.0
